Bug 1752638 - Part 0: inline ForkServiceChild::OnMessageReceived. r?#ipc-reviewers
Incoming messages to the ForkServiceChild are replies to its synchronous
requests to the fork server, read with blocking I/O, so handling the
reply can be done inline.  And there will soon be several requests,
with different replies, so it doesn't make sense to have a single
OnMessageReceived method like this.

Differential Revision: https://phabricator.services.mozilla.com/D240712