Bug 1752638 - Part 3: delay fork server shutdown until browser shutdown. r?#ipc-reviewers
Because the fork server is used to remotely `waitpid` child processes,
and this can be done until very late in shutdown, this patch extends the
fork server's lifetime.  This is a little complicated, because destroying a
`GeckoChildProcessHost` requires the IPC I/O thread to be acceptaing
tasks, and this is later than that.  So, this adds a method to extract a
`GeckoChildProcessHost`'s process handle (on Unix, the pid) and take
responsibility for handling it, and this is done for the fork server.

Differential Revision: https://phabricator.services.mozilla.com/D240715