Bug 1965103 - Turn `security.sandbox.content.headless` into sandbox level 5. r?gcp!
This was a separate pref because it also side-effects the graphics stack
by setting `MOZ_HEADLESS` for content processes.  But, it makes more
sense to have it be part of the ordered sequence of sandbox levels (for
example, the upcoming ioctl lockdown will also break graphics drivers),
and in general it's better if the sandbox policy can be understood by
looking at only one pref instead of several.

There is *no* pref migration for older profiles.  If anyone had
`security.sandbox.content.headless` set to false: if the reason was to
also turn off `webgl.out-of-process`, try resetting that pref and see if
everything still works.  If that's not the case, or if there are still
bugs with default settings, please report them, and the sandbox level
can be lowered to 4 as a workaround.

Differential Revision: https://phabricator.services.mozilla.com/D249016