Bug 1984256 - Remove the SANDBOX_HAS_USER_NAMESPACES telemetry probe. r?gcp
The probe expires in 145 and I don't expect it to tell us anything in
the foreseeable future that we haven't already learned from it.

We can re-add it in the future if need be (but ideally as a scalar,
which is what it always should have been, instead of a histogram, which
is what it was due to an accident of history).

Differential Revision: https://phabricator.services.mozilla.com/D264158